pyproject.toml
[tool.poetry]
name = "Telegram-bot"
version = "0.1.0"
description = ""
authors = ["artyombn <balabashinan@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.12.0"
apscheduler = "^3.10.4"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.33"
aiosqlite = "^0.20.0"
greenlet = "^3.0.3"
asyncpg = "^0.29.0"
pandas = "^2.2.2"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
